Moves Car and Rectangle class to external header files.
